--- src/core/services/mirrornode/schema.ts
@@ -1,33 +1,0 @@
-import type { TokenInfo } from './types';
-
-import { z } from 'zod';
-
-const mirrorKeyObject = z.object({
-  _type: z.string(),
-  key: z.string(),
-});
-
-const optionalKeyRef = z.union([mirrorKeyObject, z.null()]).optional();
-
-export const TokenInfoSchema: z.ZodType<TokenInfo> = z.object({
-  token_id: z.string(),
-  symbol: z.string(),
-  name: z.string(),
-  decimals: z.string(),
-  total_supply: z.string(),
-  max_supply: z.string(),
-  type: z.string(),
-  treasury_account_id: z.string(),
-  admin_key: optionalKeyRef,
-  kyc_key: optionalKeyRef,
-  freeze_key: optionalKeyRef,
-  wipe_key: optionalKeyRef,
-  supply_key: optionalKeyRef,
-  fee_schedule_key: optionalKeyRef,
-  pause_key: optionalKeyRef,
-  created_timestamp: z.string(),
-  deleted: z.boolean().nullable().optional(),
-  freeze_default: z.boolean().optional(),
-  pause_status: z.string(),
-  memo: z.string(),
-});
